Refactoring CDT logic to make it extensible
