--- src/client/Microsoft.Identity.Client/AuthScheme/IMsalCryptoProvider.cs
@@ -1,19 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Microsoft.Identity.Client.AuthScheme
-{
-    /// <summary>
-    /// 
-    /// </summary>
-    public interface IMsalCryptoProvider
-    {
-
-    }
-}
