+++ src/client/Microsoft.Identity.Client/AuthScheme/PoP/InMemoryCryptoProvider.cs
@@ -3,17 +3,14 @@
 
 using System;
 using System.Security.Cryptography;
-using Microsoft.Identity.Client.AuthScheme.CDT;
-using Microsoft.Identity.Client.AuthScheme.PoP;
 using Microsoft.Identity.Client.Utils;
 
-namespace Microsoft.Identity.Client.AuthScheme
+namespace Microsoft.Identity.Client.AuthScheme.PoP
 {
-
     /// <summary>
     /// The default implementation will store a key in memory    
     /// </summary>
-    internal class InMemoryCryptoProvider : IPoPCryptoProvider, ICdtCryptoProvider
+    internal class InMemoryCryptoProvider : IPoPCryptoProvider
     {
         internal /* internal for test only */ const int RsaKeySize = 2048;
         private RSA _signingKey;
