+++ src/client/Microsoft.Identity.Client/Internal/Constants.cs
@@ -37,10 +37,6 @@ namespace Microsoft.Identity.Client.Internal
         public const string PoPAuthHeaderPrefix = "PoP"; 
         public const string RequestConfirmation = "req_cnf";
         public const string BearerAuthHeaderPrefix = "Bearer";
-        public const string NoAlgorythmPrefix = "none";
-        public const string JasonWebTokenType = "JWT";
-        public const string CdtEncryptedAlgoryth = "dir";
-        public const string CdtEncryptedValue = "A256CBC-HS256";
 
         public const string ManagedIdentityClientId = "client_id";
         public const string ManagedIdentityObjectId = "object_id";
@@ -48,8 +44,6 @@ namespace Microsoft.Identity.Client.Internal
         public const string ManagedIdentityDefaultClientId = "system_assigned_managed_identity";
         public const string ManagedIdentityDefaultTenant = "managed_identity";
         public const string CiamAuthorityHostSuffix = ".ciamlogin.com";
-
-        public const string CdtRequestConfirmation = "req_ds_cnf";
 
         public static string FormatEnterpriseRegistrationOnPremiseUri(string domain)
         {
