+++ src/client/Microsoft.Identity.Client/Internal/JsonWebTokenConstants.cs
@@ -27,8 +27,5 @@ namespace Microsoft.Identity.Client.Internal
         public const string X509CertificateThumbprint = "x5t";
 
         public const string X509CertificatePublicCertValue = "x5c";
-
-        public const string CdtEncrypt = "enc";
-
     }
 }
