+++ src/client/Microsoft.Identity.Client/Platforms/netdesktop/NetDesktopPlatformProxy.cs
@@ -10,7 +10,6 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading.Tasks;
-using Microsoft.Identity.Client.AuthScheme;
 using Microsoft.Identity.Client.AuthScheme.PoP;
 using Microsoft.Identity.Client.Cache;
 using Microsoft.Identity.Client.Core;
@@ -240,7 +239,7 @@ namespace Microsoft.Identity.Client.Platforms.netdesktop
 
         public override IPoPCryptoProvider GetDefaultPoPCryptoProvider()
         {
-            return MsalCryptoProviderFactory.GetOrCreateProvider();
+            return PoPCryptoProviderFactory.GetOrCreateProvider();
         }
 
         public override IDeviceAuthManager CreateDeviceAuthManager()
